requirements.txt
av==12.3.0
certifi==2024.8.30
charset-normalizer==3.4.0
coloredlogs==15.0.1
ctranslate2==4.5.0
faster-whisper==1.1.0
filelock==3.16.1
flatbuffers==24.3.25
fsspec==2024.10.0
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.10
Jinja2==3.1.3
joblib==1.4.2
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.0
numpy==1.24.4
onnxruntime==1.16.3
openwakeword==0.6.0
packaging==24.2
pillow==10.4.0
pkg_resources==0.0.0
protobuf==5.28.3
PyAudio==0.2.14
pydub==0.25.1
python-engineio==3.14.2
python-socketio==4.6.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
scikit-learn==1.3.2
scipy==1.10.1
sentence-transformers==3.2.1
six==1.16.0
sympy==1.13.3
tflite-runtime==2.14.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.4.1+cpu
torchaudio==2.4.1+cpu
tqdm==4.67.0
transformers==4.46.3
typing_extensions==4.12.2
urllib3==2.2.3
